split service tests in separate files
