+++ .changesets/maint_geal_split_supergraph_service_tests.md
@@ -1,0 +1,5 @@
+### split service tests in separate files ([PR #3279](https://github.com/apollographql/router/pull/3279))
+
+Some of the tests became larger than the implementation, this will make them easier to navigate
+
+By [@Geal](https://github.com/Geal) in https://github.com/apollographql/router/pull/3279
